Remove unused code, confirm deletion, add favicon

+++ demo/src/pages/Home/index.tsx
@@ -7,20 +7,24 @@ import { CardItem } from '../../components/CardItem';
 import { Stack } from '@demo/components/Stack';
 import { history } from '@demo/utils/history';
 import templates from '@demo/store/templates';
+import template from '@demo/store/template';
 
 export default function Home() {
   const dispatch = useDispatch();
   const templatesState = useAppSelector('templates');
 
   const handleDelete = (id: number) => {
-    dispatch(templates.actions.deleteById({ id }));
-    Notification.success({
-      title: 'Success',
-      content: 'Template deleted successfully',
-    });
+    if (confirm('Are you sure you want to delete this template?')) {
+      dispatch(templates.actions.deleteById({ id }));
+      Notification.success({
+        title: 'Success',
+        content: 'Template deleted successfully',
+      });
+    }
   }
 
   useEffect(() => {
+    dispatch(template.actions.reset(undefined));
     dispatch(templates.actions.fetch(undefined));
   }, [dispatch]);
